# server/docs/openapi.yaml
openapi: 3.0.0
info:
  title: FitCraft API
  description: API for FitCraft workout generation and tracking app.
  version: 1.0.0
servers:
  - url: /api
    description: Local API server
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        '201':
          description: Created
        '400':
          description: Bad request (email already exists)
        '500':
          description: Server error
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
        '401':
          description: Invalid credentials
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT